add login handler and laod env vars
